requirements.txt
streamlit==1.37.0
requests==2.31.0
openai==1.31.0
beautifulsoup4==4.12.2
markdown2==2.4.10
python-dotenv==1.0.0
sqlalchemy==2.0.23
google-search-results==2.4.2
